// packages/main/src/types/AiButtonState.ts
/**
 * Different AiButton states.
 * @public
 */
enum AiButtonState {
	/**
	 * Generate state
	 * @public
	 */
	Generate = "Generate",

	/**
	 * Generating state
	 * @public
	 */
	Generating = "Generating",

	/**
	 * Revise state
	 * @public
	 */
	Revise = "Revise",
}

export default AiButtonState;